组件拆分, 添加当前分类下标签页阈值设置项
