+++ entrypoints/common/constants.ts
@@ -98,6 +98,7 @@ export const ENUM_SETTINGS_PROPS: EnumSettingsProps = {
   ALLOW_DUPLICATE_TABS: 'allowDuplicateTabs', // 同一个标签组中是否允许重复的标签页
   ALLOW_DUPLICATE_GROUPS: 'allowDuplicateGroups', // 同一个分类中是否允许重复的标签组
   LINK_TEMPLATE: 'linkTemplate', // 链接模板
+  TAB_COUNT_THRESHOLD: 'tabCountThreshold', // 分类中标签页超过该数量时，则右侧面板只展示单个分组
 };
 
 // tab 事件
