+++ entrypoints/common/locale/modules/home/zhCN.ts
@@ -29,6 +29,7 @@ export default {
 
   'home.tag.remove': '删除',
   'home.tag.recover': '还原',
+  'home.tag.countInfo': '当前分类统计',
 
   'home.tabGroup.add': '创建标签组',
   'home.tabGroup.remove': '删除该组',
@@ -54,7 +55,7 @@ export default {
     <li>点击打开标签组时，会在浏览器中依次打开该标签组中的标签页，如果设置了“打开标签页时自动从列表中删除”，则该标签组中的标签页会被删除。</li>
     <li>“还原”操作表示将回收站中的标签组还原到首页列表中。</li>
     <li>“升序”和“降序”功能只对当前分类中的非星标的标签组进行排序。</li>
-    <li><strong>注意</strong>：为了保证交互性能，分类中标签页数量超过一定数量时，右侧面板将只展示左侧选中的标签组，而不是分类下的所有标签组。</li>
+    <li><strong>注意</strong>：为了保证交互性能，分类中标签页数量超过设定值时，右侧面板将只展示选中的标签组，而不是分类下的所有标签组。</li>
     <li>分类/标签组/标签页支持拖拽排序，当分类/标签组处在选中状态时，可通过快捷键进行排序，快捷键映射见下表：</li>
   `,
 
@@ -62,5 +63,5 @@ export default {
   'home.moveTo.mergeLabel': '是否合并重复项',
   'home.moveTo.mergeTip.1': '1、勾选后，如果移动标签页，则与目标分组中相同 url 的标签页合并（去重）。',
   'home.moveTo.mergeTip.2': '2、勾选后，如果移动标签组，则与目标分类中同名的标签组合并，标签页去重。',
-  'home.tip.tooManyTabs': '分类中标签页数量超过预定值, 将只展示单个标签组',
+  'home.tip.tooManyTabs': '当前分类中标签页数量超过预定值, 将只展示单个标签组',
 }
