+++ entrypoints/options/home/TabGroup.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef, useState, useMemo } from 'react';
+import { useEffect, useRef, useState, useMemo, memo } from 'react';
 import {
   theme,
   message,
@@ -15,7 +15,7 @@ import {
   CloseOutlined,
 } from '@ant-design/icons';
 import { CopyToClipboard } from 'react-copy-to-clipboard';
-import { TagItem, GroupItem, TabItem } from '~/entrypoints/types';
+import { GroupItem, TabItem } from '~/entrypoints/types';
 import { StyledActionIconBtn } from '~/entrypoints/common/style/Common.styled';
 import { ENUM_COLORS, UNNAMED_GROUP } from '~/entrypoints/common/constants';
 import { useIntlUtls } from '~/entrypoints/common/hooks/global';
@@ -23,6 +23,7 @@ import { tabListUtils } from '@/entrypoints/common/storage';
 import DndComponent from '@/entrypoints/common/components/DndComponent';
 import DropComponent from '@/entrypoints/common/components/DropComponent';
 
+import { HomeContext } from './hooks/treeData';
 import EditInput from '../components/EditInput';
 import TabListItem from './TabListItem';
 import {
@@ -43,7 +44,6 @@ import useMoveTo from './hooks/moveTo';
 const dndKey = dndKeys.tabItem;
 
 type TabGroupProps = GroupItem & {
-  tagList?: TagItem[];
   refreshKey?: string;
   canDrag?: boolean;
   canDrop?: boolean;
@@ -74,8 +74,7 @@ const defaultGroupActions = [
 ];
 const defaultTabActions = ['remove', 'moveTo'];
 
-export default function TabGroup({
-  tagList,
+function TabGroup({
   refreshKey,
   groupId,
   groupName,
@@ -116,6 +115,7 @@ export default function TabGroup({
     onClose: onMoveToClose,
     moveData,
   } = useMoveTo();
+  const { treeDataHook } = useContext(HomeContext);
 
   const removeDesc = useMemo(() => {
     const typeName = $fmt(`home.tabGroup`);
@@ -176,16 +176,16 @@ export default function TabGroup({
     }
   };
 
-  // const handleScroll = (behavior: ScrollBehavior = 'instant') => {
-  //   if (selected && groupRef.current) {
-  //     // console.log('groupRef.current', groupRef.current)
-  //     const offsetTop = groupRef.current?.offsetTop || 0;
-  //     window.scrollTo({ top: offsetTop - 100, behavior });
-  //   }
-  // };
-  // useEffect(() => {
-  //   handleScroll();
-  // }, [selected]);
+  const handleScroll = (behavior: ScrollBehavior = 'instant') => {
+    if (selected && groupRef.current) {
+      // console.log('groupRef.current', groupRef.current)
+      const offsetTop = groupRef.current?.offsetTop || 0;
+      window.scrollTo({ top: offsetTop - 100, behavior });
+    }
+  };
+  useEffect(() => {
+    handleScroll();
+  }, [selected]);
 
   useEffect(() => {
     setTimeout(() => {
@@ -201,13 +201,14 @@ export default function TabGroup({
           body.scrollTo(0, groupTop + pagePaddingTop - window.innerHeight + 300);
         }
       }
-    }, 100);
+    }, 300);
   }, [refreshKey, selected]);
 
+  console.log('tabGroup Effect')
   useEffect(() => {
     setTimeout(() => {
       setRendering(false);
-    })
+    });
   }, []);
 
   function TabListMarkup({ tab, index }: { tab: TabItem; index: number }) {
@@ -461,7 +462,7 @@ export default function TabGroup({
       {moveToModalVisible && (
         <MoveToModal
           visible={moveToModalVisible}
-          listData={tagList}
+          listData={treeDataHook.tagList}
           moveData={moveData}
           onOk={(targetData) => {
             onMoveToConfirm(() => {
@@ -475,3 +476,5 @@ export default function TabGroup({
     </>
   );
 }
+
+export default memo(TabGroup)
