+++ entrypoints/options/home/constants.ts
@@ -2,8 +2,6 @@
 export const dndKeys = {
   tabItem: Symbol('dnd-tab-item'),
 }
-// 分类中标签页超过该数量时，则右侧面板只展示单个分组
-export const VIRTUAL_MAX_TAB_COUNT = 500;
 
 export default {
   name: 'option-home-constants',
