+++ entrypoints/types/global.ts
@@ -88,6 +88,7 @@ export type SettingsProps = {
   allowDuplicateGroups?: boolean; // 同一个分类中是否允许重复的标签组
   linkTemplate?: string; // 链接模板
   themeType?: ThemeTypes; // 主题类型
+  tabCountThreshold?: number; // 分类中标签页超过该数量时，则右侧面板只展示单个分组
 };
 
 export type EnumSettingsProps = {
@@ -102,6 +103,7 @@ export type EnumSettingsProps = {
   ALLOW_DUPLICATE_TABS: 'allowDuplicateTabs';
   ALLOW_DUPLICATE_GROUPS: 'allowDuplicateGroups';
   LINK_TEMPLATE: 'linkTemplate';
+  TAB_COUNT_THRESHOLD: 'tabCountThreshold';
 };
 
 export interface HotkeyOption {
